Allow specifying a protocol for each listener. Move HTTP stuff to Protocols::Http namespace.
